feat: CPU driving skill (per-trophy) and skill-scaled rubberbanding
- PlayerData gains skillLevel (0-10), persisted per cpuIndex for the whole
  trophy/session in RaceSettings (same lifecycle as the CPU display name),
  with a guarantee of 1-3 CPUs at max skill and a difficulty-scaled chance
  of the rest rolling above 5 (RaceSettings.GenerateCPUSkillBatch /
  EnsureCPUSkillPoolInitialized / GetOrAssignCPUSkill).
- CPUManager reads each bot's persisted skill from its PlayerData instead of
  a random/inspector-forced value; removed the now-redundant cpuLevel field.
- RaceDifficultyManager scales the "CPU behind the human, speed up" rubber-
  band boost by the CPU's own skill (0 = no boost, 10 = full configured
  boost); the "CPU ahead, slow down" branch is unchanged. Rubberband channel
  (PlayerController/PlayerStats) switched from int to float to carry the
  fractional scale.
- Fix PlayerUIMarkerSystem: the "opponents ranked ahead" position filter was
  also silencing the behind-camera proximity arrow, which almost always
  targets someone ranked behind us. The arrow is no longer gated by that
  filter, only by distance fade, like before the filter existed.

+++ Assets/Scripts/Game/PlayerStats.cs
@@ -236,7 +236,7 @@ public class PlayerStats : MonoBehaviour
         UpdateStats();
     }
 
-    public void OnRubberbandUpdated(int rubberbandLevel)
+    public void OnRubberbandUpdated(float rubberbandLevel)
     {
         rubberbandMaxSpeedEffect = rubberbandLevel;
